Modify the attack method of the monster class
